Move DEA and LP to lib modules
