+++ core/pom.xml
@@ -14,19 +14,6 @@
     <version>1.2.2-SNAPSHOT</version>
   </parent>
 
-  <!-- ================= FRESCO DEPENDENCIES ============================ -->
-
-  <dependencies>
-
-    <!-- Used for plaintext LP solving -->
-    <dependency>
-      <groupId>org.apache.commons</groupId>
-      <artifactId>commons-math3</artifactId>
-      <version>3.5</version>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-
 
   <!-- ================ FRESCO BUILD ========================== -->
