+++ lib/Readme.md
@@ -1,0 +1,5 @@
+
+Library
+=======
+
+This folder contains various functionality based on the FRESCO framework.
