+++ lib/dea/src/main/java/dk/alexandra/fresco/lib/statistics/CreditRater.java
@@ -6,6 +6,7 @@ import dk.alexandra.fresco.framework.builder.numeric.Comparison;
 import dk.alexandra.fresco.framework.builder.numeric.Numeric;
 import dk.alexandra.fresco.framework.builder.numeric.ProtocolBuilderNumeric;
 import dk.alexandra.fresco.framework.value.SInt;
+
 import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
