+++ lib/dea/src/test/java/dk/alexandra/fresco/lib/statistics/PlaintextDEASolver.java
@@ -1,21 +1,13 @@
 package dk.alexandra.fresco.lib.statistics;
 
 import dk.alexandra.fresco.lib.statistics.DeaSolver.AnalysisType;
-import java.math.BigInteger;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
 import org.apache.commons.math3.optim.PointValuePair;
-import org.apache.commons.math3.optim.linear.LinearConstraint;
-import org.apache.commons.math3.optim.linear.LinearConstraintSet;
-import org.apache.commons.math3.optim.linear.LinearObjectiveFunction;
-import org.apache.commons.math3.optim.linear.NonNegativeConstraint;
-import org.apache.commons.math3.optim.linear.Relationship;
-import org.apache.commons.math3.optim.linear.SimplexSolver;
+import org.apache.commons.math3.optim.linear.*;
 import org.apache.commons.math3.optim.nonlinear.scalar.GoalType;
+
+import java.math.BigInteger;
+import java.util.*;
+import java.util.Map.Entry;
 
 /**
  * Solves the DEA problem on a given plaintext dataset.
