+++ lib/lp/Readme.md
@@ -1,0 +1,3 @@
+
+Linear programming
+==================
