+++ lib/lp/src/main/java/dk/alexandra/fresco/lib/lp/BlandEnteringVariable.java
