+++ lib/lp/src/main/java/dk/alexandra/fresco/lib/lp/EnteringVariable.java
