+++ lib/lp/src/main/java/dk/alexandra/fresco/lib/lp/LPTableau.java
