+++ lib/lp/src/test/java/dk/alexandra/fresco/lib/lp/LpBuildingBlockTests.java
